chore: move resolve_external functions to fendermint_actor_machine crate
