+++ fendermint/actors/machine/src/lib.rs
@@ -8,16 +8,22 @@ use std::str::FromStr;
 
 use anyhow::anyhow;
 pub use fil_actor_adm::Kind;
+use fil_actors_runtime::runtime::builtins::Type;
 use fil_actors_runtime::{
-    actor_error, runtime::Runtime, ActorError, ADM_ACTOR_ADDR, FIRST_EXPORTED_METHOD_NUMBER,
-    INIT_ACTOR_ADDR,
+    actor_error, deserialize_block, runtime::Runtime, ActorError, AsActorError, ADM_ACTOR_ADDR,
+    FIRST_EXPORTED_METHOD_NUMBER, INIT_ACTOR_ADDR,
 };
 use fvm_ipld_blockstore::Blockstore;
 use fvm_ipld_encoding::ipld_block::IpldBlock;
 use fvm_ipld_encoding::tuple::*;
+use fvm_shared::error::ExitCode;
+use fvm_shared::sys::SendFlags;
 pub use fvm_shared::METHOD_CONSTRUCTOR;
 use fvm_shared::{address::Address, MethodNum};
+use num_traits::Zero;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
+
+mod ext;
 
 /// Params for creating a machine.
 #[derive(Debug, Serialize_tuple, Deserialize_tuple)]
@@ -74,6 +80,118 @@ pub const INIT_METHOD: MethodNum = 2;
 pub const GET_ADDRESS_METHOD: MethodNum = frc42_dispatch::method_hash!("GetAddress");
 /// Get machine metadata method number.
 pub const GET_METADATA_METHOD: MethodNum = frc42_dispatch::method_hash!("GetMetadata");
+
+pub enum ActorType {
+    Account,
+    EthAccount,
+    Evm,
+    Machine,
+}
+
+/// Resolve robust address and ensure it is not a Machine actor type.
+/// See `resolve_external`.
+pub fn resolve_external_non_machine(
+    rt: &impl Runtime,
+    address: Address,
+) -> Result<Address, ActorError> {
+    let (address, actor_type) = resolve_external(rt, address)?;
+    if matches!(actor_type, ActorType::Machine) {
+        Err(ActorError::illegal_argument(format!(
+            "address {} cannot be a machine",
+            address
+        )))
+    } else {
+        Ok(address)
+    }
+}
+
+/// Resolves robust address of an actor.
+pub fn resolve_external(
+    rt: &impl Runtime,
+    address: Address,
+) -> Result<(Address, ActorType), ActorError> {
+    let actor_id = rt
+        .resolve_address(&address)
+        .ok_or(ActorError::not_found(format!(
+            "actor {} not found",
+            address
+        )))?;
+    let code_cid = rt
+        .get_actor_code_cid(&actor_id)
+        .expect("failed to lookup caller code");
+    match rt.resolve_builtin_actor_type(&code_cid) {
+        Some(Type::Account) => {
+            let result = rt
+                .send(
+                    &address,
+                    ext::account::PUBKEY_ADDRESS_METHOD,
+                    None,
+                    Zero::zero(),
+                    None,
+                    SendFlags::READ_ONLY,
+                )
+                .context_code(
+                    ExitCode::USR_ASSERTION_FAILED,
+                    "account failed to return its key address",
+                )?;
+            if !result.exit_code.is_success() {
+                return Err(ActorError::checked(
+                    result.exit_code,
+                    "failed to retrieve account robust address".to_string(),
+                    None,
+                ));
+            }
+            let robust_addr: Address = deserialize_block(result.return_data)?;
+            Ok((robust_addr, ActorType::Account))
+        }
+        Some(Type::EthAccount) => {
+            let delegated_addr =
+                rt.lookup_delegated_address(actor_id)
+                    .ok_or(ActorError::forbidden(format!(
+                        "actor {} does not have delegated address",
+                        actor_id
+                    )))?;
+            Ok((delegated_addr, ActorType::EthAccount))
+        }
+        Some(Type::EVM) => {
+            let delegated_addr =
+                rt.lookup_delegated_address(actor_id)
+                    .ok_or(ActorError::forbidden(format!(
+                        "actor {} does not have delegated address",
+                        actor_id
+                    )))?;
+            Ok((delegated_addr, ActorType::Evm))
+        }
+        Some(t) => Err(ActorError::forbidden(format!(
+            "disallowed caller type {} for address {}",
+            t.name(),
+            address
+        ))),
+        None => {
+            // The caller might be a machine
+            let result = rt
+                .send(
+                    &address,
+                    GET_ADDRESS_METHOD,
+                    None,
+                    Zero::zero(),
+                    None,
+                    SendFlags::READ_ONLY,
+                )
+                .context_code(
+                    ExitCode::USR_ASSERTION_FAILED,
+                    "machine failed to return its key address",
+                )?;
+            if !result.exit_code.is_success() {
+                return Err(ActorError::forbidden(format!(
+                    "disallowed caller code {code_cid}"
+                )));
+            }
+            let robust_addr: Address = deserialize_block(result.return_data)?;
+            Ok((robust_addr, ActorType::Machine))
+        }
+    }
+}
 
 // TODO: Add method for changing owner from ADM actor.
 pub trait MachineActor {
